instrumentation library only CLEAN if forced YOU HAVE TO CALL ant build-instrumentation-clean to rebuild the library , so in case a ant set-debug is called and then a ant buil you will end HAVING NON OBFUSCATED, WITH DEBUG SYMBOL AND LOG within library

+++ instrumentation/adbi/hijack/build.sh
@@ -19,10 +19,11 @@
 
 # -d specify where to place the compiled class
 # --classpath path to externa jar and class
-
+deflib="hijack"
 debug=""
 echo "args $@"
-while getopts ":dDP" opt; do
+CLEAN=0
+while getopts ":dDPc" opt; do
   case $opt in
     P) pie="PIE=1"
       echo "position Indipendent code ENABLED"
@@ -30,18 +31,36 @@ while getopts ":dDP" opt; do
     d) debug="DEBUG=1"
       echo "debug is on"
       ;;
+    c) CLEAN=1
+      echo "force clean is on"
+      ;;
     D) echo "debug is off"
       ;;
     \?) echo "Invalid option -$OPTARG"
       ;;
   esac
 done
-
+if [ $CLEAN -eq 0 ]
+then
+  echo "checking $(pwd)/libs/*/$deflib"
+  if [ -e $(pwd)/libs/*/$deflib ]
+  then
+    echo "nothing to do ... $deflib already present"
+    exit 0
+  fi
+fi
 cdir=$(pwd)
-
-echo cleaning libs..
-rm -r ./obj/* ./libs/*
+if [ $CLEAN -eq 1 ]
+then
+  echo cleaning libs..
+  rm -r ./obj/* ./libs/*
+fi
 ndk_cmd="ndk-build V=1 $debug $pie"
 echo "Building hijack $ndk_cmd"
-cd jni &&  ndk-build V=1 clean && $ndk_cmd
+cd jni
+if [ $CLEAN -eq 1 ]
+then
+  ndk-build V=1 clean
+fi
+$ndk_cmd
 cd ${cdir}
