+++ instrumentation/adbi/instruments/binder_call/build.sh
@@ -17,8 +17,10 @@
 #      REVISION:  ---
 #===============================================================================
 debug=""
+deflib="libbhelp.so"
 echo "args $@"
-while getopts ":dDP" opt; do
+CLEAN=0
+while getopts ":dDPc" opt; do
   case $opt in
     P) pie="PIE=1"
       echo "position Indipendent code ENABLED"
@@ -26,26 +28,50 @@ while getopts ":dDP" opt; do
     d) debug="DEBUG=1"
       echo "debug is on"
       ;;
+    c) CLEAN=1
+      echo "force clean is on"
+      ;;
     D) echo "debug is off"
       ;;
     \?) echo "Invalid option -$OPTARG"
       ;;
   esac
 done
-
+if [ $CLEAN -eq 0 ]
+then
+  echo "checking $(pwd)/libs/*/$deflib"
+  if [ -e $(pwd)/libs/*/$deflib ]
+  then
+    echo "nothing to do ... $deflib already present"
+    exit 0
+  fi
+fi
 cdir=$(pwd)
 cd ../base/
-echo cleaning libs..
-rm -r ./obj/* ./libs/*
+if [ $CLEAN -eq 1 ]
+then
+  echo cleaning libs..
+  rm -r ./obj/* ./libs/*
+fi
 ndk_cmd="ndk-build V=1 $debug $pie"
 echo "Building base $ndk_cmd"
-cd jni &&  ndk-build V=1 clean && $ndk_cmd
+cd jni
+if [ $CLEAN -eq 1 ]
+then
+  ndk-build V=1 clean
+fi
+$ndk_cmd
 cd ${cdir}
 
 echo cleaning libs..
 rm -r ./obj/* ./libs/*
 ndk_cmd="ndk-build V=1 $debug $pie"
 echo "Building hijack $ndk_cmd"
-cd jni &&  ndk-build V=1 clean && $ndk_cmd
+cd jni
+if [ $CLEAN -eq 1 ]
+then
+  ndk-build V=1 clean
+fi
+$ndk_cmd
 cd ${cdir}
